Refactor route name
