+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.8.5+20",
+  "version": "0.8.5+21",
   "private": false,
   "scripts": {
     "dev": "next dev",
