+++ src/locales/cn/common.json
@@ -495,7 +495,7 @@
   "TOAST": {
     "UPLOAD_FILE_ERROR": "上传文件: <{{name}}> 时出现错误"
   },
-  "SELECT_ROLE": {
+  "CREATE_ROLE": {
     "SELECT_ROLE_ISUNFA": "选择角色 - iSunFA"
   },
