+++ src/locales/tw/common.json
@@ -495,7 +495,7 @@
   "TOAST": {
     "UPLOAD_FILE_ERROR": "上傳文件: <{{name}}> 時發生錯誤"
   },
-  "SELECT_ROLE": {
+  "CREATE_ROLE": {
     "SELECT_ROLE_ISUNFA": "選擇角色 - iSunFA"
   },
